Move non entity logic to seperate class
