+++ homeassistant/components/tahoma/tahoma_entity.py
@@ -1,14 +1,15 @@
 """Parent class for every Somfy TaHoma device."""
 import logging
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
-from pyhoma.models import Command, Device
+from pyhoma.models import Device
 
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
 from .coordinator import TahomaDataUpdateCoordinator
+from .overkiz_executor import OverkizExecutor
 
 ATTR_RSSI_LEVEL = "rssi_level"
 
@@ -28,6 +29,7 @@ class TahomaEntity(CoordinatorEntity, Entity):
         """Initialize the device."""
         super().__init__(coordinator)
         self.device_url = device_url
+        self.executor = OverkizExecutor(coordinator, device_url)
 
     @property
     def device(self) -> Device:
@@ -57,8 +59,10 @@ class TahomaEntity(CoordinatorEntity, Entity):
     @property
     def device_info(self) -> Dict[str, Any]:
         """Return device registry information for this entity."""
-        manufacturer = self.select_state(CORE_MANUFACTURER_NAME_STATE) or "Somfy"
-        model = self.select_state(CORE_MODEL_STATE) or self.device.widget
+        manufacturer = (
+            self.executor.select_state(CORE_MANUFACTURER_NAME_STATE) or "Somfy"
+        )
+        model = self.executor.select_state(CORE_MODEL_STATE) or self.device.widget
 
         return {
             "identifiers": {(DOMAIN, self.unique_id)},
@@ -67,53 +71,3 @@ class TahomaEntity(CoordinatorEntity, Entity):
             "model": model,
             "sw_version": self.device.controllable_name,
         }
-
-    def select_command(self, *commands: str) -> Optional[str]:
-        """Select first existing command in a list of commands."""
-        existing_commands = self.device.definition.commands
-        return next((c for c in commands if c in existing_commands), None)
-
-    def has_command(self, *commands: str) -> bool:
-        """Return True if a command exists in a list of commands."""
-        return self.select_command(*commands) is not None
-
-    def select_state(self, *states) -> Optional[str]:
-        """Select first existing active state in a list of states."""
-        if self.device.states:
-            return next(
-                (
-                    state.value
-                    for state in self.device.states
-                    if state.name in list(states)
-                ),
-                None,
-            )
-        return None
-
-    def has_state(self, *states: str) -> bool:
-        """Return True if a state exists in self."""
-        return self.select_state(*states) is not None
-
-    async def async_execute_command(self, command_name: str, *args: Any):
-        """Execute device command in async context."""
-        try:
-            exec_id = await self.coordinator.client.execute_command(
-                self.device.deviceurl,
-                Command(command_name, list(args)),
-                "Home Assistant",
-            )
-        except Exception as exception:  # pylint: disable=broad-except
-            _LOGGER.error(exception)
-            return
-
-        # ExecutionRegisteredEvent doesn't contain the deviceurl, thus we need to register it here
-        self.coordinator.executions[exec_id] = {
-            "deviceurl": self.device.deviceurl,
-            "command_name": command_name,
-        }
-
-        await self.coordinator.async_refresh()
-
-    async def async_cancel_command(self, exec_id: str):
-        """Cancel device command in async context."""
-        await self.coordinator.client.cancel_command(exec_id)
